Modularizacion de codigo en modulos, funciones y colores
