+++ Modulos/colores.py
@@ -1,0 +1,20 @@
+###CONSTANTES DE COLORES
+RED = '\033[31m'
+GREEN = '\033[32m'
+PURPLE = '\033[0;95m'
+RESET = '\033[0m'
+#COLORES CON NEGRITA
+B_RED = "\033[1;31m"  # Red
+B_GREEN = "\033[1;32m"  # Green
+B_YELLOW = "\033[1;33m"  # Yellow
+B_BLUE = "\033[1;34m"  # Blue
+B_PURPLE = "\033[1;35m"  # Purple
+B_CYAN = "\033[1;36m"  # Cyan
+#COLORES CON SUBRAYADO
+URED = "\033[4;31m"  # Red
+UGREEN = "\033[4;32m"  # Green
+UYELLOW = "\033[4;33m"  # Yellow
+UBLUE = "\033[4;34m"  # Blue
+#COLORES CON BACKGROUND
+BG_CYAN = "\033[0;106m"  # Cyan
+####FIN DE CONSTANTES#####
